Updated delete functionality

+++ routes/api/noteRoutes.js
@@ -71,11 +71,22 @@ router.put('/:id', async (req, res) => {
 // DELETE /api/notes/:id - Delete a note
 router.delete('/:id', async (req, res) => {
   try {
+
+    if(!req.user) {
+      return res.status(401).json({ message: "You must be logged in to delete this note." });
+    }
+
     // This needs an authorization check
     const note = await Note.findByIdAndDelete(req.params.id);
     if (!note) {
       return res.status(404).json({ message: 'No note found with this id!' });
     }
+    
+    if (note.user.toString() !== req.user._id.toString()) {
+      return res.status(403).json({ message: 'User is not authorized to delete this note.' });
+    }
+    await note.deleteOne();
+
     res.json({ message: 'Note deleted!' });
   } catch (err) {
     res.status(500).json(err);
